% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rgb.R
\name{set_alpha}
\alias{set_alpha}
\title{Set alpha value of integer colors}
\usage{
set_alpha(icol, alpha)
}
\arguments{
\item{icol}{Integer vector (where the 4 bytes making up each integer represent 
the RGBA values of the color)}

\item{alpha}{numeric alpha value in range [0, 1]. Length of 1, or same length as 'col'}
}
\value{
integer vector of colors with adjusted alpha channel
}
\description{
Set alpha value of integer colors
}
\examples{
icol <- col_to_int('red')
icol
int_to_col(icol)
icol2 <- set_alpha(icol, 0.5)
int_to_col(icol2)
}
